+++ cmd/infrared/main.go
@@ -1,17 +1,13 @@
 package main
 
 import (
-	"errors"
-	"io"
 	"log"
 	"os"
 	"os/signal"
 	"syscall"
 
-	"github.com/haveachin/infrared/configs"
 	ir "github.com/haveachin/infrared/pkg/infrared"
 	"github.com/spf13/pflag"
-	"gopkg.in/yaml.v3"
 )
 
 const (
@@ -46,14 +42,14 @@ func init() {
 }
 
 func main() {
+	log.Println("Starting Infrared")
+
 	if err := run(); err != nil {
 		log.Fatal(err)
 	}
 }
 
 func run() error {
-	log.Println("Starting Infrared")
-
 	cfg, err := loadConfig()
 	if err != nil {
 		return err
@@ -81,36 +77,4 @@ func run() error {
 	}
 
 	return nil
-}
-
-func loadConfig() (ir.Config, error) {
-	var f io.ReadCloser
-	f, err := os.Open(configPath)
-	if errors.Is(err, os.ErrNotExist) {
-		return createDefaultConfigFile()
-	} else if err != nil {
-		return ir.Config{}, err
-	}
-	defer f.Close()
-
-	cfg := ir.DefaultConfig()
-	if err := yaml.NewDecoder(f).Decode(&cfg); err != nil {
-		return ir.Config{}, err
-	}
-
-	return cfg, nil
-}
-
-func createDefaultConfigFile() (ir.Config, error) {
-	bb := configs.DefaultInfraredConfig
-	if err := os.WriteFile(configPath, bb, 0666); err != nil {
-		return ir.Config{}, err
-	}
-
-	cfg := ir.DefaultConfig()
-	if err := yaml.Unmarshal(bb, &cfg); err != nil {
-		return ir.Config{}, err
-	}
-
-	return cfg, nil
 }
