+++ configs/proxy.yml
@@ -1,0 +1,15 @@
+# This is the domain that players enter in their game client.
+# You can have multiple domains here or just one.
+# Currently this holds just a wildcard character as a domain
+# meaning that is accepts every domain that a player uses.
+# Supports '*' and '?' wildcards in the pattern string.
+#
+domains:
+  - "*"
+
+addresses:
+  - example.com:25565
+
+# Send a Proxy Protocol v2 Header to the server to forward the player IP address
+#
+#sendProxyProtocol: true
